rename plexscoring -> scoring
