+++ alphadia/workflow/peptidecentric/fragment_requantification_handler.py
@@ -9,9 +9,9 @@ from alphadia.constants.keys import CalibCols
 from alphadia.fragcomp.utils import add_frag_start_stop_idx, candidate_hash
 from alphadia.raw_data import DiaData
 from alphadia.reporting.reporting import Pipeline
-from alphadia.search.plexscoring.config import CandidateConfig
-from alphadia.search.plexscoring.plexscoring import CandidateScoring
-from alphadia.search.plexscoring.utils import (
+from alphadia.search.scoring.config import CandidateConfig
+from alphadia.search.scoring.plexscoring import CandidateScoring
+from alphadia.search.scoring.utils import (
     candidate_features_to_candidates,
 )
 from alphadia.workflow.config import Config
