+++ ruff-lint-strict.toml
@@ -9,19 +9,12 @@ exclude = [
     "alphadia/__*__.py",
     "alphadia/cli.py",
     "alphadia/exceptions.py",
-    "alphadia/plexscoring/features/**.py",
-    "alphadia/grouping.py",
-    "alphadia/outputaccumulator.py",
     "alphadia/outputtransform/**.py",
-    "alphadia/plexscoring/*.py",
-    "alphadia/quadrupole.py",
     "alphadia/search_plan.py",
     "alphadia/libtransform/**.py",
     "alphadia/search_step.py",
     "alphadia/utils.py",
     "alphadia/validate.py",
-    "alphadia/consensus/*.py",
-    "alphadia/constants/*.py",
     "alphadia/raw_data/alpharaw_wrapper.py",
     "alphadia/raw_data/bruker.py",
     "alphadia/raw_data/utils.py",
@@ -29,7 +22,6 @@ exclude = [
     "alphadia/_fdrx/stats.py",
     "alphadia/_fdrx/plotting.py",
     "alphadia/_fdrx/base.py",
-    "alphadia/numba/*.py",
     "alphadia/search/**.py",
     "alphadia/plotting/*.py",
     "alphadia/transferlearning/*.py",
